requirements.txt
Django==1.7.2
Pillow==2.7.0
django-crispy-forms==1.4.0
django-registration-redux==1.1
django-simple-captcha==0.4.3
six==1.9.0
wsgiref==0.1.2
